git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/tiger2checker@14881 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/arg/ARGMergeJoinPredicatedAnalysis.java
@@ -38,11 +38,13 @@ import org.sosy_lab.cpachecker.exceptions.PredicatedAnalysisPropertyViolationExc
 
 public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
 
+  private final boolean deleteSubgraphAfterMerge;
   private final MergeOperator wrappedMerge;
-  private ArrayList<ARGState> toDeleteFromReached = new ArrayList<>();
+  private final ArrayList<ARGState> toDeleteFromReached = new ArrayList<>();
 
-  public ARGMergeJoinPredicatedAnalysis(MergeOperator pWrappedMerge) {
+  public ARGMergeJoinPredicatedAnalysis(MergeOperator pWrappedMerge, final boolean pDeleteSubgraph) {
     wrappedMerge = pWrappedMerge;
+    deleteSubgraphAfterMerge = pDeleteSubgraph;
   }
 
   // may cause problems during refinement, relink of elements may lead to non-matching abstraction formulae
@@ -74,9 +76,13 @@ public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
     if (retElement.equals(wrappedState2)) { return pState2; }
 
     ARGState mergedElement = new ARGState(retElement, null);
-    // now replace argElement2 by mergedElement in ARG
-    // deleteChildren(argElement2);
-    deleteChildren2(argElement2);
+
+    if (deleteSubgraphAfterMerge) {
+      // now replace argElement2 by mergedElement in ARG
+      // deleteChildren(argElement2);
+      deleteChildren2(argElement2);
+    }
+
     argElement2.replaceInARGWith(mergedElement);
 
 
@@ -194,12 +200,12 @@ public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
 
       // stop deletion and find out if there are deeper children which are and may reference one of their ancestor
       toProcess.addAll(laterCovered);
-      while(!toProcess.isEmpty()){
+      while (!toProcess.isEmpty()) {
         current = toProcess.pop();
 
-        for(ARGState c: current.getChildren()) {
+        for (ARGState c: current.getChildren()) {
 
-          assert(c.getParents().size()==1);
+          assert (c.getParents().size()==1);
           // relink or delete child
           if (c.getCoveredByThis().size() != 0) {
             // relink child in ARG to parent of first covered element
@@ -228,14 +234,14 @@ public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
 
       // find out if now covered by external node
       boolean changed = true;
-      while(changed){
+      while (changed) {
         changed = false;
-        for(ARGState later:laterCovered){
-          if(later.getCoveredByThis().size()!=0){
+        for (ARGState later:laterCovered) {
+          if (later.getCoveredByThis().size()!=0) {
             covered = getCoveredNodeFromDifferentSubtree(subtreeNodes, later);
             if (covered != null) {
               // delete edge from parent and introduce covering
-              assert(later.getParents().size()<=1);
+              assert (later.getParents().size()<=1);
               if (later.getParents().size() == 1) {
                 (new ARGState(later.getWrappedState(), later.getParents().iterator().next())).setCovered(later);
                 later.getParents().iterator().next().deleteChild(later);
@@ -256,7 +262,7 @@ public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
 
       // delete rest of subtree
       toProcess.addAll(laterCovered);
-      while(!toProcess.isEmpty()){
+      while (!toProcess.isEmpty()) {
         current = toProcess.pop();
         toDeleteFromReached.add(current);
 
@@ -275,7 +281,7 @@ public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
     ARGState covered;
     do {
       covered = coveredElems.next();
-      assert(covered.getCoveredByThis().size()==0);
+      assert (covered.getCoveredByThis().size()==0);
       if (covered.getParents().size() == 0 || subtreeNodes.contains(covered)) {
         covered = null;
       }
@@ -283,17 +289,17 @@ public class ARGMergeJoinPredicatedAnalysis implements MergeOperator {
     return covered;
   }
 
-  private HashSet<ARGState> getSubtreeNodes(ARGState top){
+  private HashSet<ARGState> getSubtreeNodes(ARGState top) {
     Stack<ARGState> toProcess = new Stack<>();
     HashSet<ARGState> nodes = new HashSet<>();
 
     toProcess.push(top);
     nodes.add(top);
 
-    while(!toProcess.isEmpty()){
+    while (!toProcess.isEmpty()) {
       top = toProcess.pop();
-      for(ARGState child:top.getChildren()){
-        if(nodes.add(child)){
+      for (ARGState child:top.getChildren()) {
+        if (nodes.add(child)) {
           toProcess.push(child);
         }
       }
